Refactor CompressionArg into compression.rs
